Changes in the LogIn page(WIP) and add redux folders and store
